Inline autoalloc fragment into its screen
